% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{createPlateInput}
\alias{createPlateInput}
\title{Create plate annotation file}
\usage{
createPlateInput(rawDir, file, entries = c(), batchAsFolder = FALSE,
  csvFormat = "csv2")
}
\arguments{
\item{rawDir}{A character string specifying the directory of the raw data folder}

\item{file}{The output file name and path for the plate annotation file}

\item{entries}{A vector of characters you wish to use as the column names in the plate annotation file (optional)}

\item{batchAsFolder}{A logical value. Whether the raw data from different batches are stored in different subfolders. Default is FALSE.}

\item{csvFormat}{A character sting indicating the format of the csv file, either "csv" (comma separated) or "csv2" (semicolon separated). Default is "csv2".}
}
\description{
This function creates a csv file required for annotating plates, based on the raw data files
}
