% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{createWellInput}
\alias{createWellInput}
\title{Create well annotation file}
\usage{
createWellInput(file, colNum, rowNum, entries = c("name",
  "concentration"), platePerSample = 1, csvFormat = "csv2")
}
\arguments{
\item{file}{The output file name and directory for the sample annotation file}

\item{colNum}{The number of columns on the plate, will be labelled by number}

\item{rowNum}{The number of rows on the plate, will be labelled by alphabets}

\item{entries}{A vector of characters you wish to use as the column names in the drug file (optional)}

\item{platePerSample}{The number of plates used for each sample in the screen. Default is 1.}
}
\description{
This function creates a csv file required for annotating wells on drug screening plate
}
