% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{detectClass}
\alias{detectClass}
\title{Detect data type, used for shiny app
A funtion to detect wether a vector is numeric, binary or catagorical.}
\usage{
detectClass(x)
}
\arguments{
\item{x}{A vector of input data}
}
\description{
Detect data type, used for shiny app
A funtion to detect wether a vector is numeric, binary or catagorical.
}
