% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{fitEdgeEffect}
\alias{fitEdgeEffect}
\title{Fit a 2D surface to describe the edge effect on each plate}
\usage{
fitEdgeEffect(screenData, method = "loess", useNeg = TRUE,
  useLowConcentrations = 0, span = 1, exclude = c())
}
\arguments{
\item{screenData}{The tidytable containing screening results generated by readScreen() function}

\item{method}{A character string specifing the method used for estimating edge effect. Currently two methods are supported: "loess", which uses local regression to estimate edge effect; "sigmoid", a experimental feature, using 2D-sigmoid model to estimate edge effect.}

\item{useNeg}{A logical value. If TRUE, only negative controls are used to estimate edge effect. Otherwise, all the wells are used for edge effect estimation.}

\item{useLowConcentrations}{A integer value. In addition to negative controls, sample wells with low concentrations of drugs or other perturbators can be considered as "de facto" negative controlls and included in edge effect estimation. The default value is 0, means no sample wells are used. If the value n >= 1, the n lowest concentrations will be used.}

\item{span}{A numeric value. The span parameter for loess, which controls the degree of smoothing.}

\item{exclude}{A list of samples that should be excluded when performing edge effect correction. Both plate file names or sampleIDs can be used as identifiers.}
}
\description{
This function fits a 2D surface using either loess model or 2D sigmod model based on the measurements on each plate to describe the edge effect.
}
