% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{fitIC50}
\alias{fitIC50}
\title{Use the dr4pl package to perform IC50 fit. Can be directly used for geom_smooth() in ggplot2}
\usage{
fitIC50(formula, data = NULL, weights = NULL, logDose = NULL, ...)
}
\arguments{
\item{formula}{Formula for the curve fitting.}

\item{data}{A data frame contain the raw concentration and the viability value. The viability should not be the percent viability value.}

\item{...}{Parameters passed to logLogisticRegression()}

\item{weigths}{Not used, mainly for geom_smooth() purpose}
}
\description{
Use the dr4pl package to perform IC50 fit. Can be directly used for geom_smooth() in ggplot2
}
