% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReport.R
\name{makeReport}
\alias{makeReport}
\title{Generate a quality report of the screen data}
\usage{
makeReport(
  screenData,
  showCode = FALSE,
  title = "Report for my drug screening project",
  author = "Author Name",
  ifPlatePlot = FALSE
)
}
\arguments{
\item{screenData}{a data table that contains the screen results generated by \code{readScreen()} function.}

\item{showCode}{a boolean value, whether to show code in the html report.}

\item{title}{a character string specifying the title for the html report}

\item{author}{a character string specifying the author name}

\item{ifPlatePlot}{a boolean value, whether to make plate heatmap plots in the report.
Plate heatmap plots can be time consuming for large screen datasets.}
}
\value{
This function creates an html file named 'report.html' in a folder named 'report' in the working directory.
}
\description{
This function generates a customized Rmarkdown based html report for priliminary quality assurance.
}
\examples{
# load processed data
data(screenData_normalized)

# make html report
makeReport(screenData_normalized, showCode = TRUE,
           title = 'Report example',
           author = 'Author Name', ifPlatePlot = FALSE)

}
