% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{makeShiny}
\alias{makeShiny}
\title{Generate interactive shiny app
A funtion to generate customized shiny app for interactive exploration of the screen data}
\usage{
makeShiny(screenData, sampleAnnotations = c("batch", "sampleID", "patientID"))
}
\arguments{
\item{screenData}{The tidytable containing screening results generated by readScreen() function}

\item{sampleAnnotations}{A vector of character strings, specifying the columns that should be used as sample annotations.}
}
\description{
Generate interactive shiny app
A funtion to generate customized shiny app for interactive exploration of the screen data
}
