% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QA.R
\name{platePlot}
\alias{platePlot}
\title{Plot the layout or measurements on each screening plate}
\usage{
platePlot(screenData, plotPlate = "all", plotType = "layout",
  ifCorrected = FALSE, limits = NULL, pdfName = NULL, ncol = 2,
  nrow = 3, width = 16, height = 16)
}
\arguments{
\item{screenData}{The tidytable containing screening results generated by readScreen() function}

\item{plotPlate}{A character string or a vector of character strings specifying the file name of the plates. If all plates should be plotted, "all" can be used. Default value is "all"}

\item{plotType}{A character string, either "layout", "zscore", "viability" can be used. If "layout" is specified, the plate layout will be plotted in a heatmap according to the information in the "wellType" column. If "zscore" is chosen, the zscore of each well on a plate will be plotted in the heatmaps. If "viability" is used, the viability value on each plate will be plotted. This choise is only valied when the percent inhibition values were calculated when read in the screen data or using the normalizePlate function. Default is "layout"}

\item{ifCorrected}{A logical string, whether to plot the orignial viablity values or viablity values after edge effect correction. The default value is FALSE, plot the original value.}

\item{limits}{A numeric vector indicates the lower and upper limits of the value shown in the heatmap. The default values is (0,2) for plotType= "inhibition" and (-3,3) for plotType = "zscore".}

\item{pdfName}{A character string specifying the output pdf file name. If not specified, a list containing the ggplot objects of plate plots will be returned. Default is NULL.}

\item{ncol}{The number of columns in a page. Default number is 2.}

\item{nrow}{The number of rows in a page. Default number is 3.}

\item{width}{The width of the page}

\item{height}{The height of the page}
}
\description{
This function plot a heatmap or a series of heatmaps showing the plate layout or the measurements on the screening plates
}
