% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPlate.R
\name{plotPlate}
\alias{plotPlate}
\title{Plot the layout or measurements on each screening plate}
\usage{
plotPlate(
  screenData,
  plotPlate = "all",
  plotType = "layout",
  ifCorrected = FALSE,
  limits = NULL,
  pdfName = NULL,
  ncol = 2,
  nrow = 3,
  width = 16,
  height = 16
)
}
\arguments{
\item{screenData}{a data frame containing screen data generated by \code{readScreen()} function}

\item{plotPlate}{a character string or a vector of character strings specifying the file name of the plates. If all plates should be included, 'all' can be used. Default value is 'all'}

\item{plotType}{a character string specifying the type of the plot, one of 'layout', 'zscore' and 'viability'.
If 'layout' is specified, the plate layout will be plotted in a heatmap according to the information in the 'wellType' column.
If 'zscore' is chosen, the zscore of each well on a plate will be plotted in the heatmaps.
If 'viability' is used, the normalized intensity or viability values on each plate will be plotted.
This choice is only valid when plate normalization is performed by \code{readScreen()} function or \code{normalizePlate()} function. Default is 'layout'}

\item{ifCorrected}{a logical value, whether to plot values after edge effect correction. The default value is FALSE, plot the original value.}

\item{limits}{a numeric vector indicates the lower and upper limits of the value shown in the heatmap.
The default values is \code{[0,2]} for \code{plotType = "viability"} and \code{[-3,3]} for \code{plotType = 'zscore'}.}

\item{pdfName}{a character string specifying the output pdf file name. If not specified, a list containing the ggplot objects of plate plots will be returned. Default is NULL.}

\item{ncol}{the number of columns in a pdf page. Default number is 2.}

\item{nrow}{the number of rows in a pdf page. Default number is 3.}

\item{width}{the width of the page, in points}

\item{height}{the height of the page, in points}
}
\value{
If the argument \code{pdfName = NULL}, this function returns a list of ggplot objects and each object is a heatmap for each plate. If a file name is specified \
for the \code{pdfName} argument, a pdf file with the specified name and layout will be created in the working directory.
}
\description{
This function generates a heatmap or a series of heatmaps showing the plate layout, raw intensity or normalized intensity for each plate in the screen dataset.
}
\examples{
# load processed data
data('screenData_normalized')

# plot the cell viability heatmap on all plates
plotPlate(screenData_normalized, plotPlate = 'all', plotType = 'viability')
# Please see the vignette for more information.
}
