% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotRawCount.R
\name{plotRawCount}
\alias{plotRawCount}
\title{Plot the distribution of the raw signal intensity (raw counts) on each plate}
\usage{
plotRawCount(screenData, topN = NULL, ifLog10 = FALSE)
}
\arguments{
\item{screenData}{a data frame containing screen date generated by readScreen() function}

\item{topN}{an integer value or NULL, indicating how many plates should be included in the plot If NULL, all the plates are included}

\item{ifLog10}{a logical value, whether to perform log10 transformation}
}
\value{
a ggplot object
}
\description{
This function generates a box plot to show the raw signal distribution on each plate.
}
\examples{
# load processed data
data('screenData')

# plot the raw ATP counts on all plates
plotRawCount(screenData)
# Please see the vignette for more information.
}
