% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QA.R
\name{plotTypeDist}
\alias{plotTypeDist}
\title{Plotting raw signal distribution for each well type
A function to plot the raw signal distribution in annotated well types on a single plate.}
\usage{
plotTypeDist(screenData, plotPlate = "all", ifLog10 = FALSE,
  pdfName = NULL)
}
\arguments{
\item{screenData}{The tidytable containing screening results generated by readScreen() function}

\item{plotPlate}{A character string or a vector of character strings specifying the file name of the plates. If all plates should be plotted, "all" can be used. Default value is "all"}

\item{ifLog10}{A logical value, whether to perform log10 transformation on raw data}

\item{pdfName}{A character string specifying the output pdf file name. If not specified, a list containing the ggplot objects of plate plots will be returned. Default is NULL.}
}
\description{
Plotting raw signal distribution for each well type
A function to plot the raw signal distribution in annotated well types on a single plate.
}
