% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addons.R, R/processData.R
\name{predict.fitIC50}
\alias{predict.fitIC50}
\title{Generic function for ic50 class generated from fitIC50 function.}
\usage{
\method{predict}{fitIC50}(object, newdata = NULL, se.fit = FALSE,
  level = 0.95, interval = c("none", "confidence", "prediction"), ...)

\method{predict}{fitIC50}(object, newdata = NULL, se.fit = FALSE,
  level = 0.95, interval = c("none", "confidence", "prediction"), ...)
}
\arguments{
\item{object}{Object of class inheriting from "fitIC50"}

\item{newdata}{An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.}

\item{se.fit}{Not used, mainly for geom_smooth purpose}

\item{level}{Not used, mainly for geom_smooth purpose}

\item{interval}{Not used, mainly for geom_smooth purpose}
}
\description{
Generic function for IC50 class generated from fitIC50 function.
}
