% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{screenData}
\alias{screenData}
\title{Example drug screen data on CLL samples}
\format{
an object of "tbl_df" (tidy table)
}
\usage{
data(screenData)
}
\description{
This is an example drug screen dataset of CLL samples processed by \code{readScreen()} function.
Normalization was not performed.
}
\examples{
data(screenData)
}
\keyword{datasets}
