% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{screenData_normalized}
\alias{screenData_normalized}
\title{Example drug screen data on CLL samples with normalization}
\format{
an object of "tbl_df" (tidy table)
}
\usage{
data(screenData_normalized)
}
\description{
This is an example drug screen dataset of CLL samples processed by \code{readScreen()} function.
Normalization was performed based on negative controls.
}
\examples{
data(screenData_normalized)
}
\keyword{datasets}
