% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumIC50.R
\name{sumIC50}
\alias{sumIC50}
\title{Summarise drug effect across concentrations using a 4-parameter log-logistic model}
\usage{
sumIC50(formula, data = NULL, minConc = 0, maxConc = 15, n = 100, ...)
}
\arguments{
\item{formula}{formula for the curve fitting.}

\item{data}{a data frame containing the raw concentration and the viability value.}

\item{minConc}{lower bound of concentrations for the integration of area under curve, default value is 0.}

\item{maxConc}{upper bound of concentrations for the integration of area under curve, default value is 15.}

\item{n}{number of bins for integration.}

\item{...}{further arguments to be passed to dr4pl function.}
}
\value{
a dataframe with only one row and four columns:
\item{UpperLimit}{the upper limit of the fitted logistic model}
\item{IC50}{the half maximal inhibitory concentration (IC50),
or the half maximal effective concentration (EC50), depends on the type of input data}
\item{Slope}{the slope of the fitted logistic model}
\item{lowerLimit}{the upper limit of the fitted logistic model}
\item{AUC}{the area under the fitted sigmoid doese-response curve}
If the model fitting is failed, all values in the dataframe will be NA.
}
\description{
This function will fit a 4-parameter log-logistic model using dr4pl package based on the input dose-response table and return the model parameters as well as integrated area under the fitted sigmoid curve.
}
\examples{
# create an example dose-response table
doseTab <- data.frame(viability = c(1.0, 0.95, 0.5, 0.2, 0.1),
                      concentration = c(0.001, 0.01, 0.1, 1, 10))

# get the parameters
sumIC50(viability ~ concentration, data = doseTab)
}
