% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{correctEdgeEffect}
\alias{correctEdgeEffect}
\title{Correct edge effect for each screen plate}
\usage{
correctEdgeEffect(screenData, estimateMethod = "loess", useNeg = TRUE,
  useLowConcentrations = 0, span = 1, correctMethod = "bliss",
  exclude = c())
}
\arguments{
\item{screenData}{The tidytable containing screening results generated by readScreen() function}

\item{useNeg}{A logical value. If TRUE, only negative controls are used to estimate edge effect. Otherwise, all the wells are used for edge effect estimation.}

\item{useLowConcentrations}{A integer value. In addition to negative controls, sample wells with low concentrations of drugs or other perturbators can be considered as "de facto" negative controlls and included in edge effect estimation. The default value is 0, means no sample wells are used. If the value n >= 1, the n lowest concentrations will be used.}

\item{span}{A numeric value. The span parameter for loess, which controls the degree of smoothing.}

\item{correctMethod}{A character string specifying the method used for subtracting the edge effect. Currently two methods "bliss" and "linear" are supported.}

\item{exclude}{A list of samples that should be excluded when performing edge effect correction.}

\item{method}{A character string specifing the method used for estimating edge effect. Currently two methods are supported: "loess", which uses local regression to estimate edge effect; "sigmoid", a experimental feature, using 2D-sigmoid model to estimate edge effect.}
}
\description{
This function subtract the incubation (or edge) effect from the observed signal intensity to retain the true signal intensity. If the incubation effect has already been estimated by fitEdgeEffect() function, the values in the edgeFactor column will be used. Otherwsie, this function will call fitEdgeEffect function to calculate the edge effect.
}
