% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPlateInput.R
\name{createPlateInput}
\alias{createPlateInput}
\title{Create a template for plate annotation file}
\usage{
createPlateInput(rawDir, file, entries = c(), batchAsFolder = FALSE)
}
\arguments{
\item{rawDir}{a character string specifying the directory of the raw data folder}

\item{file}{the output file name and path for the plate annotation file}

\item{entries}{a vector of character strings to use as the column names in the plate annotation file (optional)}

\item{batchAsFolder}{a logical value, whether the raw data from different batches are stored in different sub-folders.
The default value is \code{FALSE}. If \code{TRUE}, the folder names will be used as batch identifiers automatically.}
}
\value{
This function creates a tab-separated values (tsv) file in the working directory.
}
\description{
This function creates a tab-separated values (tsv) file required for annotating plates, for example, the sample ID or patient ID for each plate.
Users can use the file created by this function as a starting point for preparing the plate annotation file required by the \code{readScreen()} function.
Users can also prepare the plate annotation file from scratch, as long as it follows the format requirement. More information is available in the package vignette.
}
\examples{
# create a template for plate annotation file
rawFolder <- system.file('testData/rawData', package = 'DrugScreenExplorer')
createPlateInput(rawDir = rawFolder, file = 'plateAnno.tsv',
    entries = c('sampleID', 'patientID'))
# see vignette for more instructions
}
