% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QA.R
\name{estimateEdgeEffect}
\alias{estimateEdgeEffect}
\title{Function for estimating edge effect on each plate}
\usage{
estimateEdgeEffect(screenData, nLayer = 1, onlyNeg = TRUE,
  identifier = "fileName", pdfName = NULL)
}
\arguments{
\item{screenData}{The tidytable containing screening results generated by readScreen() function}

\item{nLayer}{An integer value larger than 1, indicating up to how many layers on the edge of the screen are considered as edges. The default value is 1.}

\item{onlyNeg}{A logical value, whether only well annotated as negative controls are used for edge effect estimation. If FALSE, all the measurements on each plate are used for edge effect estimation. The functions will automatically check wehther there are negative control wells on the edge or inside the well under the partition of edge wells and inner wells specified by nLayer parameter. Default is TRUE.}

\item{identifier}{A character string specifing the colname in the screenData table that will be used as identifiers in the plot. Default value is fileName. If one identifier matches several plates, the error bar will be shown in the plot.}

\item{pdfName}{Either NULL or a character string specifying the output file name for the plot in pdb format. If NULL, no plot will be created.}
}
\description{
This function estimates the intensity of edge effect on each screen plate by comparing the median measurment values on the edge and on the inner plate. And plot the edge effect factor distributions among plates.
}
