% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitIC50.R
\name{fitIC50}
\alias{fitIC50}
\title{Use the dr4pl package to perform IC50 curve fitting. Can be directly used for geom_smooth() in ggplot2.}
\usage{
fitIC50(formula, data = NULL, weights = NULL, logDose = NULL, ...)
}
\arguments{
\item{formula}{formula for the curve fitting.}

\item{data}{a data frame that contains the  concentration values and the normalized viability values or percent inhibition values.}

\item{weights}{not used, mainly for geom_smooth() purpose}

\item{logDose}{a numeric value specifying the base of log transformation for concentration,
if the concentration is log transformed. The default value is NULL (not log transformed).}

\item{...}{further arguments to be passed to dr4pl function}
}
\value{
fitIC50 returns an object of class 'fitIC50', which is a list containing the following components:
\item{model}{the input dose-response table}
\item{formula}{the formula used for the model fitting}
\item{parm_fit}{the fitted parameters for a 4-paramater logistic model}
\item{logDose}{a numeric value specifying the base of log transformation for concentration,
if the concentration is log transformed. The default value is NULL (not log transformed).}
}
\description{
Use the dr4pl package to perform IC50 curve fitting. Can be directly used for geom_smooth() in ggplot2.
}
\examples{
# create an example dose-response table
doseTab <- data.frame(viability = c(1.0, 0.95, 0.5, 0.2, 0.1),
                      concentration = c(0.001, 0.01, 0.1, 1, 10))

# fit a four-parameter logistic model using dr4pl package
fitIC50(viability ~ concentration, data = doseTab)

}
