% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{makeReport}
\alias{makeReport}
\title{Generate Rmarkdown report}
\usage{
makeReport(screenData, showCode = FALSE,
  title = "Report for my drug screening project",
  author = "John Smith", ifPlatePlot = FALSE)
}
\arguments{
\item{screenData}{The tidytable containing screening results generated by readScreen() function.}

\item{showCode}{Boolean value, whether to show code in the html.}

\item{title}{Character string specifying the title for the html report}

\item{author}{Character string specifying the author name}

\item{ifPlatePlot}{Boolean value, whether to make plate plots in the report. Plate plots for large screens are time consuming.}
}
\description{
A function to generate customized html Rmarkdown report for quality assessment
}
