% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeShiny.R
\name{makeShiny}
\alias{makeShiny}
\title{Generate a shiny app to explore process data interactively}
\usage{
makeShiny(screenData, sampleAnnotations = c("sampleID", "patientID"))
}
\arguments{
\item{screenData}{a data frame that contains the processed screen data generated by \code{readScreen()} function}

\item{sampleAnnotations}{a vector of character strings, specifying the columns that should be used as sample annotations.
The information in those columns can be used to label samples in visualization or to perform association test.}
}
\value{
This function creates a folder named 'shiny' in the working direction.
In the 'shiny' folder, 'app.R' is the R script for the shiny app and 'shinyData.RData' contains the processed data for the shiny app.
}
\description{
This function generates a customized shiny app for interactive exploration of the processed and annotated screen data
}
\examples{
# load processed data
data(screenData_normalized)

# make html report
makeShiny(screenData_normalized)

}
