% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{normalizePlate}
\alias{normalizePlate}
\title{Per-plate normalization}
\usage{
normalizePlate(screenData, method = "negatives", discardLayer = 0)
}
\arguments{
\item{screenData}{A tidytable that contains the output from a readScreen function.}

\item{method}{A character string, either "negatives" or "both", specifying the method for calculating percent inhibition values. If "negatives" is used, each measurement on the plate will be divided by the median of the values of negative control wells on the same plate. If "both" is used, a "normalized percent inhbition (NPI)" is applied in a per-plate basis. For an inhibition assay, this method divides the difference between each measurement on a plate and the median measurement of the positive controls on that plate by the difference between the median mesurement of negative controls and positive controls on that well. The default value is "negatives".}

\item{discardLayer}{An integer values. To avoid the impact of incubation effect on the negative controls, a certain number of the exterior on the plate can be ignored when summarising the measurement of controls. 0 means do not discard any control wells.}
}
\description{
This function calculates the percent inhibition values for each plate based on the control wells on the same plate.
}
