% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTypeDist.R
\name{plotTypeDist}
\alias{plotTypeDist}
\title{Plotting raw signal distribution for each well type}
\usage{
plotTypeDist(screenData, plotPlate = "all", ifLog10 = FALSE, pdfName = NULL)
}
\arguments{
\item{screenData}{a data frame containing the screen data generated by readScreen() function}

\item{plotPlate}{a character string or a vector of character strings specifying the file name of the plates to be included in the plots.
If all plates should be included, 'all' can be used. The default value is 'all'.}

\item{ifLog10}{a logical value, whether to perform a log10 transformation on the values.}

\item{pdfName}{a character string specifying the output pdf file name.
If not specified, a list containing the ggplot objects will be returned. Default is NULL.}
}
\value{
If the argument \code{pdfName = NULL}, this function returns a list of ggplot objects. If a file name is specified \
for the \code{pdfName} argument, a pdf file that contains the plots will be created in the working directory.
}
\description{
A function to generate box plots to show the raw signal distribution for each well type on all or selected plates.
For this function to work, the "wellType" column must be present in the data frame.
}
\examples{
# load processed data
data('screenData')

# plot raw ATP count distribution for each well type
plotTypeDist(screenData)
# Please see the vignette for more information.
}
