% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictIC50.R
\name{predict.fitIC50}
\alias{predict.fitIC50}
\title{Predict responses based on input dose information and fitted log-logistic model}
\usage{
\method{predict}{fitIC50}(
  object,
  newdata = NULL,
  se.fit = FALSE,
  level = 0.95,
  interval = c("none", "confidence", "prediction"),
  ...
)
}
\arguments{
\item{object}{an object of class 'fitIC50', generated by fitIC50 function}

\item{newdata}{a numeric vector or data fram with one column that contains the new concentration values
If omitted, the concentrations used to fit the model are used.}

\item{se.fit}{not used, mainly for geom_smooth purpose}

\item{level}{not used, mainly for geom_smooth purpose}

\item{interval}{not used, mainly for geom_smooth purpose}

\item{...}{further arguments to be passed to ggplot when using geom_smooth().}
}
\value{
a vector of predicted response based on a fitted model and input concentration values.
}
\description{
This function predicts the response based on input fitted log-logistic model and concentration values.
}
\examples{
# create an example dose-response table
doseTab <- data.frame(viability = c(1.0, 0.95, 0.5, 0.2, 0.1),
                      concentration = c(0.001, 0.01, 0.1, 1, 10))

# fit a four-parameter logistic model using dr4pl package
mm <- fitIC50(viability ~ concentration, data = doseTab)

# Predict response using the input concentrations
predict(mm)

# Predict response using new concentrations
predict(mm, newdata = c(0.25, 2.5, 5, 15))
}
