% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importData.R
\name{readScreen}
\alias{readScreen}
\title{Read the screening results from a whole experiment}
\usage{
readScreen(rawDir, plateAnnotationFile, wellAnnotationFile, rowRange, colRange,
  sep, negWell = c(), posWell = c(), normalization = FALSE,
  method = "negatives", discardLayer = 0, commaAsDecimal = FALSE,
  csvFormat = "csv2")
}
\arguments{
\item{rawDir}{A character string specifying the directory of the raw data folder.}

\item{plateAnnotationFile}{A character string specifying the plate annotation file.}

\item{wellAnnotationFile}{A character string specifying the well annotation file.}

\item{rowRange}{Either An integer vector with length of two or an integer value. If it's an vector, the range of the lines that contain the result values are speficified. Otherwise, only the starting line is indicated.}

\item{colRange}{Either An integer vector with length of two or an integer value. If it's an vector, the range of the columns (separated by delimiters) that contain the result values are speficified. Otherwise, only the starting column is indicated.}

\item{sep}{A character string specifying the delimiters for extracting values in the text file. A regular expression can be used. Default value is using both tab and semicolon as delimiter.}

\item{negWell}{A character vector or a character string of well IDs or drug names (if the "name" column is present in well annotation), which specifies the negative control wells}

\item{posWell}{A character vector or a character string of well IDs or drug names (if the "name" column is present in well annotation), which specifies the positive control wells}

\item{normalization}{A logical value, whether to calculate percent inhibition based on control wells on the plate. Default is FALSE}

\item{method}{A character string, either "negatives" or "both", specifying the method for calculating percent inhibition values. If "negatives" is used, each measurement on the plate will be divided by the median of the values of negative control wells on the same plate. If "both" is used, a "normalized percent inhbition (NPI)" is applied in a per-plate basis. For an inhibition assay, this method divides the difference between each measurement on a plate and the median measurement of the positive controls on that plate by the difference between the median mesurement of negative controls and positive controls on that well. The default value is "negatives".}

\item{discardLayer}{An integer values. To avoid the impact of incubation effect on the negative controls, a certain number of the exterior on the plate can be ignored when summarising the measurement of controls. 0 means do not discard any control wells.}

\item{commaAsDecimal}{A logical value, whether commas is used as decimal in the screen value. Default if FALSE.}

\item{csvFormat}{A character string specifying the csv file type, either "csv" or "csv2". Default is csv2}
}
\description{
This function creates a tidy table containing the screening result from a whole experiment. Drugs and samples are annotated based on drug annotation and sample annotation file
}
