% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{summariseScreen}
\alias{summariseScreen}
\title{Summarise the drug effect across concentrations}
\usage{
summariseScreen(screenData, method = "average")
}
\arguments{
\item{screenData}{The tidytable containing screening results generated by readScreen() function}

\item{method}{A character string specifing the method used for summarising the effect. Currently three methods are supported: average, which simply calculates the mean effect under all concentrations; AUC, which calculates the normalized area under dose-reponse curve using linear-log trapezoidal method; IC50, which fits a robust four parameter log-logistic model using dr4pl package.}
}
\description{
This function summarises the normalized drug effect (viability) across multiple concentrations into a single value. The summarisation will be based on normalized viability values, so plate normalisation needs to be performed first. If the edge effect corrected viabiliy values are present, the normalisation will be performed on both uncorrected and corrected viabilities. A suffix, ".cor" will be added to the name of summarised values.
}
